requirements.txt
aniso8601==7.0.0
argcomplete==1.9.3
attrs==19.1.0
beautifulsoup4==4.8.1
bleach==3.1.0
boto3==1.9.156
botocore==1.12.253
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
ChocoPY==0.1.3
Click==7.0
docutils==0.15.2
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-REST==1.3
Flask-RESTful==0.3.7
Flask-SocketIO==4.2.1
gevent==1.4.0
greenlet==0.4.15
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
MarkupSafe==1.1.1
mongoengine==0.18.2
pkginfo==1.5.0.1
pycparser==2.19
Pygments==2.4.2
PyJWT==1.7.1
pymongo==3.9.0
python-dateutil==2.8.0
python-engineio==3.9.3
python-socketio==4.3.1
pytz==2019.2
readme-renderer==24.0
redis==3.3.8
requests==2.22.0
requests-toolbelt==0.9.1
s3transfer==0.2.1
six==1.12.0
socketIO-client==0.7.2
soupsieve==1.9.5
tqdm==4.38.0
twine==2.0.0
urllib3==1.25.6
webencodings==0.5.1
websocket-client==0.56.0
Werkzeug==0.15.6
